\name{Rz-package}
\docType{package}
\alias{Rz-package}
\alias{Rz}
\title{GUI Tool for Data Management like SPSS or Stata}
\description{
R is very powerfull but not enough for social science becouse of lack
in some functionality for managing survey data SPSS or Stata provides.
Memisc package supplements those, variable labels, value labels,
definable missing values and so on, but to efficiently work these functions
need graphical interface to overlook data. This package
provides such graphical interface similar to SPSS's Variable View and
data managing system using memisc package as backend. Additionally, Rz has powerful
plot assistant interface based on ggplot2.
}
\details{
Simply excute Rz() to start.
}

\keyword{ memisc }
