\name{rzAddData}
\alias{rzAddData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Add data.set to Rz
}
\description{
Add data.set to Rz
}
\usage{
rzAddData(data.set, name = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data.set}{
data.set or data.frame to be added. 
}
  \item{name}{
data.set.name
}
}
\details{
Add data.set to Rz
}
\author{
Masahiro Hayashi
}
\seealso{
\code{\link{rzReloadData}}
}
