\name{RzThemeEditor}
\alias{RzThemeEditor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Theme Editor for ggplot2
}
\description{
Theme Editor for ggplot2
}
\usage{
RzThemeEditor(new = FALSE, ...)
}
\arguments{
  \item{new}{
logical: rebuild UI
}
  \item{\dots}{
currently no use
}
}
\details{
Start RzThemeEditor.
}
\author{
Masahiro Hayashi
}
\seealso{
\code{\link{Rz}}%% ~~objects to See Also as \code{\link{help}}, ~~~
}
